requirements.txt
python-vlc
Pillow